Reduce size for types

+++ scraper/src/models/active_match.rs
@@ -6,19 +6,19 @@ use serde::{Deserialize, Serialize};
 pub struct ActiveMatch {
     pub start_time: u32,
     pub winning_team: u8,
-    pub match_id: u64,
+    pub match_id: u32,
     pub players: Vec<ActiveMatchPlayer>,
     pub lobby_id: u64,
     pub net_worth_team_0: u32,
     pub net_worth_team_1: u32,
     pub duration_s: u32, // Currently always 0
-    pub spectators: u32,
-    pub open_spectator_slots: u32,
+    pub spectators: u16,
+    pub open_spectator_slots: u16,
     pub objectives_mask_team0: u16,
     pub objectives_mask_team1: u16,
     pub match_mode: MatchMode,
     pub game_mode: GameMode,
-    pub match_score: u32,
+    pub match_score: u16,
     pub region_mode: RegionMode,
 }
 
@@ -34,7 +34,7 @@ pub struct ActiveMatchPlayer {
 pub struct ClickHouseActiveMatch {
     pub start_time: u32,
     pub winning_team: u8,
-    pub match_id: u64,
+    pub match_id: u32,
     #[serde(rename = "players.account_id")]
     pub players_account_id: Vec<u64>,
     #[serde(rename = "players.team")]
@@ -47,13 +47,13 @@ pub struct ClickHouseActiveMatch {
     pub net_worth_team_0: u32,
     pub net_worth_team_1: u32,
     pub duration_s: u32, // Currently always 0
-    pub spectators: u32,
-    pub open_spectator_slots: u32,
+    pub spectators: u16,
+    pub open_spectator_slots: u16,
     pub objectives_mask_team0: u16,
     pub objectives_mask_team1: u16,
     pub match_mode: MatchMode,
     pub game_mode: GameMode,
-    pub match_score: u32,
+    pub match_score: u16,
     pub region_mode: RegionMode,
 }
